sun_stage/pyproject.toml
[tool.poetry]
name = "sun-stage"
version = "0.1.0"
description = "Management of Internship App"
authors = ["Stephanie-siewe <siewe.stephanie@gmail.com>"]
readme = "README.md"
packages = [{include = "sun_stage"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic = {extras = ["email"], version = "^2.5.2"}
sqlalchemy = "^2.0.23"
pyotp = "^2.9.0"
alembic = "^1.12.1"
aiosqlite = "^0.19.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
google-api-python-client = "^2.110.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[scripts]
start = "python app.py"